refactor: fix to return earlier
